#93: [Refactor] アニメーション状態管理ロジックをカスタムフックに分離
